infrastructure/functions/authorizer/requirements.txt
pyjwt[crypto]>=2.0.0